Move types to eolearn.core.types (#534)
* move types to `eolearn.core.types`

* move parser types to eolearn.core.types

* fix types in conftest

* fix import of type alias on newer python version

* delay annotations to enable docstring type alises
