+++ docs/source/conf.py
@@ -73,8 +73,8 @@ extensions = [
 # Include typehints in descriptions
 autodoc_typehints = "description"
 autodoc_type_aliases = {
-    "FeaturesSpecification": "eolearn.core.parsing.FeaturesSpecification",
-    "SingleFeatureSpec": "eolearn.core.parsing.SingleFeatureSpec",
+    "FeaturesSpecification": "eolearn.core.types.FeaturesSpecification",
+    "SingleFeatureSpec": "eolearn.core.types.SingleFeatureSpec",
 }
 
 # Both the class’ and the __init__ method’s docstring are concatenated and inserted.
