+++ features/eolearn/features/blob.py
@@ -8,6 +8,7 @@ Copyright (c) 2017-2022 Matej Aleksandrov, Devis Peressutti, Žiga Lukšič (Sin
 This source code is licensed under the MIT license found in the LICENSE
 file in the root directory of this source tree.
 """
+from __future__ import annotations
 
 from math import sqrt
 from typing import Any, Callable
@@ -16,7 +17,7 @@ import numpy as np
 import skimage.feature
 
 from eolearn.core import EOPatch, EOTask
-from eolearn.core.utils.parsing import SingleFeatureSpec
+from eolearn.core.types import SingleFeatureSpec
 
 
 class BlobTask(EOTask):
@@ -109,7 +110,7 @@ class DoGBlobTask(BlobTask):
         max_sigma: float = 30,
         threshold: float = 0.1,
         overlap: float = 0.5,
-        **kwargs: Any
+        **kwargs: Any,
     ):
         super().__init__(
             feature,
@@ -119,7 +120,7 @@ class DoGBlobTask(BlobTask):
             max_sigma=max_sigma,
             threshold=threshold,
             overlap=overlap,
-            **kwargs
+            **kwargs,
         )
 
 
@@ -136,7 +137,7 @@ class DoHBlobTask(BlobTask):
         max_sigma: float = 30,
         threshold: float = 0.1,
         overlap: float = 0.5,
-        **kwargs: Any
+        **kwargs: Any,
     ):
         super().__init__(
             feature,
@@ -147,7 +148,7 @@ class DoHBlobTask(BlobTask):
             max_sigma=max_sigma,
             threshold=threshold,
             overlap=overlap,
-            **kwargs
+            **kwargs,
         )
 
 
@@ -164,7 +165,7 @@ class LoGBlobTask(BlobTask):
         max_sigma: float = 30,
         threshold: float = 0.1,
         overlap: float = 0.5,
-        **kwargs: Any
+        **kwargs: Any,
     ):
         super().__init__(
             feature,
@@ -175,5 +176,5 @@ class LoGBlobTask(BlobTask):
             max_sigma=max_sigma,
             threshold=threshold,
             overlap=overlap,
-            **kwargs
+            **kwargs,
         )
