+++ visualization/eolearn/visualization/eopatch.py
@@ -8,6 +8,8 @@ Copyright (c) 2017-2022 Žiga Lukšič, Devis Peressutti, Nejc Vesel, Jovan Viš
 This source code is licensed under the MIT license found in the LICENSE
 file in the root directory of this source tree.
 """
+from __future__ import annotations
+
 import datetime as dt
 import itertools as it
 from dataclasses import dataclass, field
@@ -20,7 +22,7 @@ from geopandas import GeoDataFrame
 from pyproj import CRS
 
 from eolearn.core import EOPatch, FeatureType
-from eolearn.core.utils.parsing import SingleFeatureSpec
+from eolearn.core.types import SingleFeatureSpec
 
 from .eopatch_base import BaseEOPatchVisualization, BasePlotConfig
