+++ visualization/eolearn/visualization/eopatch_base.py
@@ -8,6 +8,8 @@ Copyright (c) 2017-2022 Žiga Lukšič, Devis Peressutti, Nejc Vesel, Jovan Viš
 This source code is licensed under the MIT license found in the LICENSE
 file in the root directory of this source tree.
 """
+from __future__ import annotations
+
 import abc
 import datetime as dt
 from dataclasses import dataclass
@@ -17,8 +19,9 @@ import numpy as np
 from geopandas import GeoDataFrame
 
 from eolearn.core import EOPatch
+from eolearn.core.types import SingleFeatureSpec
 from eolearn.core.utils.common import is_discrete_type
-from eolearn.core.utils.parsing import SingleFeatureSpec, parse_feature
+from eolearn.core.utils.parsing import parse_feature
 
 
 @dataclass
